requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
setuptools==73.0.1
sqlparse==0.5.1
uritemplate==4.1.1
